// src/app/interfaces/cituro.ts
export interface Cituro {

    id: number;
    datum: Date;
    tag: string;
    kommentar: string;
    pilotVorname: string;
    pilotNachname: string;
    sdlVorname: string;
    sdlNachname: string;
    affVorname: string;
    affNachname: string;
    instrVorname: string;
    instrNachname: string;
    videoVorname: string;
    videoNachname: string;


}